+++ iOS/Catalog/Catalog/OptionSelection/OptionCard/OptionCardView.swift
@@ -6,59 +6,6 @@
 //
 
 import SwiftUI
-import Combine
-
-enum OptionCardModel {
-
-  struct State: Equatable {
-    var hashTag: [String]
-    var info: ModelTypeContent
-    var isModalPresenting = false
-  }
-
-  enum ViewAction {
-    case onTapDetail(isPresenting: Bool)
-  }
-
-}
-
-protocol OptionCardViewIntentType {
-  
-  var state: OptionCardModel.State { get }
-
-  func send(action: OptionCardModel.ViewAction)
-
-  func send(action: OptionCardModel.ViewAction, viewEffect: (() -> Void)?)
-  
-}
-
-final class OptionCardViewIntent: ObservableObject {
-  
-  init(initialState: State) {
-    state = initialState
-  }
-  
-  typealias State = OptionCardModel.State
-  typealias ViewAction = OptionCardModel.ViewAction
-  
-  @Published var state: State
-  
-  var cancellable: Set<AnyCancellable> = []
-  
-}
-
-extension OptionCardViewIntent: OptionCardViewIntentType, IntentType {
-  
-  func mutate(action: OptionCardModel.ViewAction, viewEffect: (() -> Void)?) {
-    switch action {
-    case .onTapDetail(let isPresenting):
-      state.isModalPresenting = isPresenting
-    }
-  }
-  
-}
-
-
 
 struct OptionCardView: IntentBindingType, View {
