+++ iOS/Catalog/Catalog/OptionSelection/OptionCardScroll/OptionCardScrollModel.swift
